pyproject.toml
[tool.poetry]
name = "poetry shell"
version = "0.1.0"
description = ""
authors = ["Vladimir Sarychev <svmyhome@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
faker = "^15.3.1"
allure-pytest = "^2.11.1"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
jsonschema = "^4.17.0"
selene = "^2.0.0b14"
curlify = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.5"
black = {extras = ["d"], version = "^22.10.0"}

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
#log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--alluredir=allure-results --clean-alluredir"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
